[RestAPI] Move MetadataDao to core folder
